Changed from solving for p to solving for pmh (static pressure minus hydrostatic pressure).
While p and pmh (pd in OpenFOAM-1.5.?) are equivalent on orthogonal meshes they
are not on non-orthogonal meshes and the difference is very important for
buoyancy-dominated flows such as settling.  settlingFoam is now written in terms
of pmh (static pressure minus hydrostatic pressure) which used to be called pd
but that confused too may people.